Submit

+++ app/Controllers/Account.php
@@ -20,7 +20,8 @@ use DateTime;
 /**
  * Account Controller
  */
-class Account extends Controller {
+class Account extends Controller
+{
 
     /**
      * Handling login action of customer. </br>
@@ -28,7 +29,8 @@ class Account extends Controller {
      * 
      * @return void
      */
-    public function loginAction() {
+    public function loginAction()
+    {
 
         $logics = new CustomerLogics();
         if (isset($_COOKIE["User"]) && $logics->search("email", $_COOKIE["User"])) {
@@ -57,7 +59,8 @@ class Account extends Controller {
      * Redirect to default page in this app if success. Else, stay in register page.
      * @return void
      */
-    public function registerAction() {
+    public function registerAction()
+    {
         if (isset($_POST["first_name"]) && isset($_POST["last_name"]) && isset($_POST["register_email"]) && isset($_POST["register_phone"])) {
             $first_name = $_POST["first_name"];
             $last_name = $_POST["last_name"];
@@ -76,10 +79,12 @@ class Account extends Controller {
      * Redirect to default page in this app if success. Else, stay in recovery page.
      * @return void
      */
-    public function recoveryAction() {
+    public function recoveryAction()
+    {
         if (!isset($_GET["remail"])) {
             View::render('recovery.php');
         }
+        
         $email = $_GET['remail'];
         $Logics = new CustomerLogics();
         $result = $Logics->recovery($email);
@@ -90,18 +95,22 @@ class Account extends Controller {
      * Accept an User has been login in the system do infor action.
      * @return type
      */
-    public function inforAction() {
+    public function inforAction()
+    {
         $logics = new CustomerLogics();
+
         if (isset($_COOKIE["User"]) && $logics->search("email", $_COOKIE["User"])) {
             $cus = $logics->search('email', $_COOKIE["User"]);
             View::render('infor.php', array('cus' => $cus));
             return;
         }
+
         if (isset($_SESSION['User']) && $logics->search("email", $_SESSION['User'])) {
             $cus = $logics->search('email', $_SESSION["User"]);
             View::render('infor.php', array('cus' => $cus));
             return;
         }
+
         View::render('login.php');
     }
 
@@ -109,7 +118,8 @@ class Account extends Controller {
      * Help user log out System. Delete all Coockie, Session related to them.
      * @return void 
      */
-    public function logoutAction() {
+    public function logoutAction()
+    {
         unset($_SESSION["User"]);
         setcookie("User", "", time() - 60, "/", "", 0);
         View::render('index.html');
@@ -119,47 +129,47 @@ class Account extends Controller {
      * Accept an User has been login in the system update his information.
      * @return type
      */
-    public function updateAction() {
+    public function updateAction()
+    {
         $logics = new CustomerLogics();
+
         if (!isset($_SESSION['User']) || !$logics->search("email", $_SESSION['User'])) {
             View::render('login.php');
             return;
         }
+
         $cus = $logics->search("email", $_SESSION['User']);
+
         if (isset($_POST["first_name"]) && isset($_POST["last_name"]) && isset($_POST["position"]) && isset($_POST["dob"]) && isset($_POST["phone"]) && isset($_POST["address"])) {
-            $params = array(
+            $params = [
                 'first_name' => $_POST["first_name"],
                 'last_name' => $_POST["last_name"],
                 'position' => $_POST["position"],
                 'birth_date' => DateTime::createFromFormat("Y-m-d", $_POST["dob"])->format('Y-m-d'),
                 'phone' => $_POST["phone"],
                 'address' => $_POST["address"]
-            );
+            ];
             $logics->update($_SESSION['User'], $params);
             return;
         }
-        if (isset($_FILES["avatar"])) {
-            $image = $_FILES['avatar'];
-            $imagetemp = $image['tmp_name'];
 
+        if (isset($_FILES["avatar"]) && is_uploaded_file($_FILES['avatar']['tmp_name'])) {
             //The path you wish to upload the image to
             $imagePath = __DIR__ . '\/../../' . 'public/asset/images/';
+            $save_path = $imagePath . "C" . $cus->getCustomerId() . "_avatar.png";
 
-            if (is_uploaded_file($imagetemp)) {
-                $save_path = $imagePath . "C" . $cus->getCustomerId() . "_avatar.png";
-                if (file_exists($save_path)) {
-                    unlink($save_path);
-                }
-                if (move_uploaded_file($imagetemp, $save_path)) {
-                    $params = array(
-                        'avatar' => "C" . $cus->getCustomerId() . "_avatar.png"
-                    );
-                    $logics->update($_SESSION['User'], $params);
-                }
+            if (file_exists($save_path)) {
+                unlink($save_path);
+            }
+
+            if (move_uploaded_file($_FILES['avatar']['tmp_name'], $save_path)) {
+                $params = [
+                    'avatar' => "C" . $cus->getCustomerId() . "_avatar.png"
+                ];
+                $logics->update($_SESSION['User'], $params);
             }
         }
         $cus = $logics->search('email', $_SESSION["User"]);
         View::render('infor.php', array('cus' => $cus));
     }
-
 }
